+++ query/src/sql/statements/mod.rs
@@ -18,7 +18,6 @@ mod analyzer_expr;
 mod analyzer_statement;
 mod analyzer_value_expr;
 mod statement_alter_user;
-mod statement_compact_table;
 mod statement_copy;
 mod statement_create_database;
 mod statement_create_stage;
@@ -33,6 +32,7 @@ mod statement_explain;
 mod statement_grant;
 mod statement_insert;
 mod statement_kill;
+mod statement_optimize_table;
 mod statement_revoke;
 mod statement_select;
 mod statement_select_convert;
@@ -55,7 +55,6 @@ pub use analyzer_statement::QueryAnalyzeState;
 pub use analyzer_statement::QueryRelation;
 pub use query::QueryASTIR;
 pub use statement_alter_user::DfAlterUser;
-pub use statement_compact_table::DfOptimizeTable;
 pub use statement_copy::DfCopy;
 pub use statement_create_database::DfCreateDatabase;
 pub use statement_create_stage::DfCreateStage;
@@ -71,6 +70,7 @@ pub use statement_grant::DfGrantObject;
 pub use statement_grant::DfGrantStatement;
 pub use statement_insert::DfInsertStatement;
 pub use statement_kill::DfKillStatement;
+pub use statement_optimize_table::DfOptimizeTable;
 pub use statement_revoke::DfRevokeStatement;
 pub use statement_select::DfQueryStatement;
 pub use statement_set_variable::DfSetVariable;
