+++ query/src/storages/fuse/table.rs
@@ -118,7 +118,7 @@ impl Table for FuseTable {
     }
 
     async fn optimize(&self, ctx: Arc<QueryContext>, keep_last_snapshot: bool) -> Result<()> {
-        self.do_truncate_history(ctx, keep_last_snapshot).await
+        self.do_optimize(ctx, keep_last_snapshot).await
     }
 }
