+++ query/tests/it/interpreters/interpreter_table_truncate.rs
