+++ query/tests/it/storages/fuse/operations/purge_drop.rs
@@ -1,0 +1,47 @@
+//  Copyright 2021 Datafuse Labs.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+use common_base::tokio;
+use common_exception::Result;
+
+use crate::storages::fuse::table_test_fixture::append_sample_data;
+use crate::storages::fuse::table_test_fixture::check_data_dir;
+use crate::storages::fuse::table_test_fixture::execute_command;
+use crate::storages::fuse::table_test_fixture::TestFixture;
+
+#[tokio::test]
+async fn test_fuse_history_truncate_in_drop_stmt() -> Result<()> {
+    let fixture = TestFixture::new().await;
+    let db = fixture.default_db_name();
+    let tbl = fixture.default_table_name();
+    let ctx = fixture.ctx();
+    fixture.create_default_table().await?;
+
+    // ingests some test data
+    append_sample_data(10, &fixture).await?;
+    // let's Drop
+    let qry = format!("drop table '{}'.'{}'", db, tbl);
+    execute_command(qry.as_str(), ctx.clone()).await?;
+    // there should be no files left inside test root (dirs are kept, though)
+    check_data_dir(
+        &fixture,
+        "drop table: there should be no file left",
+        0,
+        0,
+        0,
+    )
+    .await;
+    Ok(())
+}
