+++ tests/suites/0_stateless/09_0005_remote_insert_into_select.result
@@ -1,2 +1,1 @@
 20
-20
