+++ tests/suites/0_stateless/09_0005_remote_insert_into_select.sql
@@ -7,8 +7,6 @@ create table n1(a uint64);
 insert into n1 select number from numbers(10);
 insert into n1 select number from numbers(10);
 select count(*) from n1;
-compact table n1;
-select count(*) from n1;
 
 DROP TABLE n1;
 DROP DATABASE db_09_0005;
