+++ tests/suites/0_stateless/09_0007_func_fuse_truncate_purge.sql
