+++ tests/suites/0_stateless/09_0008_fuse_optimize_table.result
@@ -1,0 +1,15 @@
+4
+1
+5
+6
+7
+5
+6
+7
+1
+5
+6
+7
+8
+9
+10
