+++ tests/suites/0_stateless/09_0008_fuse_optimize_table.sql
@@ -1,0 +1,44 @@
+DROP DATABASE IF EXISTS db_09_0008;
+CREATE DATABASE db_09_0008;
+USE db_09_0008;
+
+create table t(a uint64);
+
+---------------------------
+
+insert into t values (5);
+insert into t values (6);
+insert into t values (7);
+
+optimize table t compact;
+
+-- expects 4 history items, 3 of previous insertion, 1 for last compaction
+select count(*) from fuse_history('db_09_0008', 't');
+
+---------------------------
+
+-- purge data and history
+optimize table 't' purge;
+-- expect 1 snapshot left
+select count(*) from fuse_history('db_09_0008', 't');
+-- check the data
+select * from t order by a;
+-- purge again, nothing happens
+optimize table 't' purge;
+select * from t order by a;
+
+---------------------------
+
+insert into t values (8);
+insert into t values (9);
+insert into t values (10);
+-- purge and compact
+optimize table 't' all;
+-- expect 1 snapshot left
+select count(*) from fuse_history('db_09_0008', 't');
+select * from t order by a;
+
+---------------------
+
+DROP TABLE t;
+DROP DATABASE db_09_0008;
